Adding DOC file about terminal colors.
This is the first doc file, so is the introduction of the doc folder
that should contain generic info and discusions. In the case of this
first file, ir contain info about the bultin color schemes in
gnome-terminal so it can be replicated in other terminal emulators. Also
some fancy color I found to use with the terminal.