Adding term-colors.sh to show terminal colors.
This script from the interwebs show all 8/16 colors in foreground
and background and the defaults fg/bg in a matrix with all the
combinations. It is useful to see if the color palette is good to
be visible in all combinations, something that sometimes is a
problem, specially with programing and comments highlight.