Adding setdesktoplayout[.c] & switch_desktops.sh.
setdesktoplayout.c is a small program to change layout of virtual
desktops in a window-managed X session. With It we can set desktops in
horizontal (default in openbox) and vertical lines (gnome shell) and in
a matrix (unity).
switch_desktops.sh is a script to swap all windows between two desktops.
It have some configuration parameters but nothing fancy.