Adding right align functionality to cascade.sh.
In cascade.sh the new '-r' option organize windows aligned with the
right border. It is useful when a window that has a smaller width comes
after to one larger, so the right corner keeps aways visible, even when
window focus of the larger one go above the it. See figure below:

NORMAL (ALIGN LEFT)       ALIGN RIGHT
+--------------+          +--------------+
|              |          |              |
|   +-------+  |          |         +-------+
|   |       |  |          |         |       |
|   |   +-----------+     |        +-----------+
|   |   |           |     |        |           |