Changes in gitconfig.
Changes in alias name from 'lol' to 'll'.
Add alias 'ls' that is like ll but list changed files for each commit.
Remove default identity (username, name, email) to help not commit in a
repo with the wrong identity (if no identity is set the commit will not
abort but will give a lengthly message on completion).