Add deskconfig.py and setdesktopname.py
setdesktopname.py lets you set the name of a virtual desktop.

deskconfig.py is a script to organize the windows in the desktop.
It is not finished yet, currently it is more of a prof of concept and
just take the windows that are closest to the corners of the desktop
and put they exactly at the corners.
In the future this script should, along with a lot of windows
organization, supersede cascade.sh