Adding support to startx and X under systemd
With systemd there is this security ting that the X session can only
start in the same tty that it is called. So we fix the code in initx.sh.
Also we add xinitrc to support "startx" command and unify initx.sh with
xinitrc. What we do is:
  * xinitrc have all the logic to setup the enviroment under X
  * initx.sh just sets up X and sources xinitrc.
With that we can use X in the same way with or without startx or xinit.