tmux.conf: adding new keybind to toggle mouse handle
now, with prefix-m and prefix-M one can enable or disable the mouse
handling within tmux. It enable you can:
- copy
- enter copy mode
- change panes
- resize panes
- change windows
The cavet is that while enabled any of this modes, the terminal emulator
can't handle the mouse events. So that is why we have the toggles.
Also add initial setup for mouse at global level. It starts disabled.