vimrc: better mouse config and vimdiff highlight
When using vim through tmux, it fail to recognize a xterm2 compatible
terminal, that has some fancy mouse support for drag and drop. So we
are hardcoding it in vimrc.
Vim have always have this awful highlight default for diffs that don't
work well with syntax highlight enabled. It would just change the
background color of some line and not the foreground, so if the colors
in the foreground were equal or close to the background you would not
be able to read what is in it. So we changed the foreground in said
lines to white. It was about time to fix it.