conky: WIP add support to new config format (awesome.conf/.lua)
Conky changed to a lua-based config format in 1.10. Add support for the
new format and a first version with dynamic configuration.

awesome.lua dynamic config: With the new lua config we can embed smarter
logic to the configuration, like discover number of ethernet interfaces
and mount points at load time. So this new config is a first step in
this direction.