xinitrc: Update wallpaper handling to use wallsetter.sh
It is a script, a first trying to set wallpapers in a random way. It is
not good, uses feh as a backend and is very slow. But it kinda work. So
using it for now and maybe take a hand at it again in the future to
improve it.