vimrc: Move viminfo place to inside of .vim
viminfo is one of this dotfiles that clutter the $HOME. Who new you can
change its place. Now it is inside of the .vim folder, nice and clean.
The only caveat it is that vim has a bug with the option and
nocompatible setting. It is fixed in the next commit, separated from
this because in the future we may want to remove it.