tmux: Change pane color for C-a to blue
It is my most used one and I am preferring a blue color for the pane
because with my urxvt pallet it is easier on the eyes and less
distracting. And also the yellow is already my vim status bar, so this
helps to not mix them visually and is more ~aesthetics~ ... maybe.

This commit was split from original
WIP/catch-all/backup/temporary commit:

  commit e07089e2758644bcccb96b79418364bc2bb88db3
  Author:     RFF <rff@email.com>
  AuthorDate: Tue Feb 14 10:55:20 2017 -0200
  Commit:     RFF <rff@email.com>
  CommitDate: Tue Feb 14 13:32:44 2017 -0200

      WIP: tmux.conf changes need refactoring

      Some of this changes were cherry picked in other branchs.
      Needs to consolidate all of this.