tmux: Change pane color for C-c to yellow (and for arrr)
I don't remember why I changed it, maybe I wanted to same magenta for
uses on the term or I wanted the yellow for arrr. Anyway, because C-c is
the default prefix for arrr, so we also change it for arrr.

This commit was split from original
WIP/catch-all/backup/temporary commit:

  commit e07089e2758644bcccb96b79418364bc2bb88db3
  Author:     RFF <rff@email.com>
  AuthorDate: Tue Feb 14 10:55:20 2017 -0200
  Commit:     RFF <rff@email.com>
  CommitDate: Tue Feb 14 13:32:44 2017 -0200

      WIP: tmux.conf changes need refactoring

      Some of this changes were cherry picked in other branchs.
      Needs to consolidate all of this.