openbox: Refactor rc.xml to not use multi-tag oneliners
Most of the file uses one tag per line and with identation for the depth
level. So this is to just make all the file the same formatting. We
loose some vertical space and easier to compare multiple keybinds of the
same group, but with xml this types of oneliners become too big anyway,
so I am thinking it is not worth it. Also This multi-line identation is
better for diff.

This commit was split from original
WIP/catch-all/backup/temporary commit:

  commit 78420e20bed51d43585ae3b8e1320a06aa5c053a
  Author:     RFF <rff@email.com>
  AuthorDate: Tue Feb 14 13:17:11 2017 -0200
  Commit:     RFF <rff@email.com>
  CommitDate: Tue Feb 14 13:32:44 2017 -0200

      WIP: openbox: rc.xml changes need refactoring

      Some of this changes were cherry picked in other branchs.
      Needs to consolidate all of this.