openbox: Change amixer volume up/down steps
On arrr, FF86Audio{Raise,Lower}Volume keys were "a touch pane", so the
sensibility and repeat of key presses where pretty high, so it made
sense at the time to have a very small step. This is no longer the case
for bisonho or any other computer still being used by me. So we can
revert to a more sensible increment for use with buttons.

For the mouse shortcuts, The inverse is true, The mouse I used to use
had this wheels with step feedback, I guess the resolution of the wheel
was small. So in any case, the logic here is the other way around.

This commit was split from original
WIP/catch-all/backup/temporary commit:

  commit 78420e20bed51d43585ae3b8e1320a06aa5c053a
  Author:     RFF <rff@email.com>
  AuthorDate: Tue Feb 14 13:17:11 2017 -0200
  Commit:     RFF <rff@email.com>
  CommitDate: Tue Feb 14 13:32:44 2017 -0200

      WIP: openbox: rc.xml changes need refactoring

      Some of this changes were cherry picked in other branchs.
      Needs to consolidate all of this.