openbox: Remove repeated client-list-combined-menu action
There was two keybinds with the same action, this is not the best way to
declare such keybinds (you can put multiples keybinds separated by space
in the key attribute), in any case The extra keybind was not really
useful or I don't remember if it was a workaround for something (maybe I
used to mixup A-Menu and W-Menu). In any case, lets remove it.

This commit was split from original
WIP/catch-all/backup/temporary commit:

  commit 78420e20bed51d43585ae3b8e1320a06aa5c053a
  Author:     RFF <rff@email.com>
  AuthorDate: Tue Feb 14 13:17:11 2017 -0200
  Commit:     RFF <rff@email.com>
  CommitDate: Tue Feb 14 13:32:44 2017 -0200

      WIP: openbox: rc.xml changes need refactoring

      Some of this changes were cherry picked in other branchs.
      Needs to consolidate all of this.