openbox: Add redundant keybind to close window
Not the right way to do it, as seen in a previews commit removing a
repeated keybind. But I guess When I first did it, maybe openbox
didn't have support for the syntax I mentioned. In any case, this is a
preview to have all WM keybinds with W-*, but for now I will not remove
the rightly standard A-F4 shortcut.

This commit was split from original
WIP/catch-all/backup/temporary commit:

  commit 78420e20bed51d43585ae3b8e1320a06aa5c053a
  Author:     RFF <rff@email.com>
  AuthorDate: Tue Feb 14 13:17:11 2017 -0200
  Commit:     RFF <rff@email.com>
  CommitDate: Tue Feb 14 13:32:44 2017 -0200

      WIP: openbox: rc.xml changes need refactoring

      Some of this changes were cherry picked in other branchs.
      Needs to consolidate all of this.