ipython: Add minimal configuration
This config changes the default behavior to:

  - Not print a initial banner message on start up
  - Not print an empty line between prompts
  - Not ask to confimation to exit

You can create a new template profile with

  ipython profile create foobar

This is the type that has all configuration options commented out and
with detailed explanation, but terrible to maintain in a repository. So
we kept it at a minimum and use

  ipython -h # and
  ipython profile -h

To see more info about it.

This commit was split from original
WIP/catch-all/backup/temporary commit:

  commit c8f6cdedf8ca1bd0e16d70ac47f196fc5a8e4f73
  Author:     RFF <rff@email.com>
  AuthorDate: Fri Dec 23 15:09:01 2016 -0200
  Commit:     RFF <rff@email.com>
  CommitDate: Tue Feb 14 13:32:44 2017 -0200

      WIP: Add IPython custom config. Need some cleaning

      This was a generated template/default and I just changed somethings.
      I think there it more than needed and maybe we can just add some small
      things instead of a full template.